requirements.txt
PyQt5
paho-mqtt
matplotlib
numpy
paramiko